renamed and moved around a few submodules for clarity
