+++ src/tests/test_load_napistu_graphs.py
@@ -6,11 +6,11 @@ import pytest
 import torch
 from torch_geometric.data import Data
 
-from napistu_torch.labeling.constants import (
+from napistu_torch.labels.constants import (
     LABEL_TYPE,
     LABELING,
 )
-from napistu_torch.labeling.labeling_manager import LabelingManager
+from napistu_torch.labels.labeling_manager import LabelingManager
 from napistu_torch.load.constants import (
     SPLITTING_STRATEGIES,
     VALID_SPLITTING_STRATEGIES,
